Fix variable grid sizing issue
The grid items had hardcoded sizes in styles.css.
They have been updated so that their sizes are
assigned dynamically through javascript.
A future update could be replacing the
e.target.setAttribute(... part on line 31 of the
js with something that won't overwrite the prexis-
ting styles on the grid boxes.